requirements.txt
flake8
lxml
matplotlib
notebook
openpyxl
pandas
numpy
pillow
plotly
python-dateutil
requests
sqlalchemy
xlrd
xlsxwriter
xlutils
xlwings
xlwt
pytrends
pyxlsb
python-dotenv
PyQt6
isort
